split magefile into a package in ./magefiles and drop unused code
