+++ .pre-commit-config.yaml
@@ -1,6 +1,7 @@
 exclude: |
   (?x)(
     (^vendor/)|
+    (^.vscode/)|
     (mage_output_file.go)|
     (.deepcopy.go$)
   )
